requirements.txt
# Flask Web Framework
Flask>=3.0.0
flask-cors>=4.0.0

# AI/ML Libraries - Version uyumluluğu için >= kullanıyoruz
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.25.0

# Image Processing
Pillow>=10.0.0

# Utilities
numpy>=1.24.0

